backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
anyio==4.4.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.47.0
tqdm==4.66.4
rapidfuzz==3.9.7
pdfplumber==0.11.4
APScheduler==3.10.4
redis==5.0.8